chore: rename app scraper utils dir and throw app-scraper-errors from utils
