+++ src/WorkerProcess/AppScraperUtils/AppScraperError.ts
@@ -1,0 +1,11 @@
+import { PushAnalyticsError } from '../../PushAnalyticsError'
+
+export class AppScraperError extends PushAnalyticsError {
+    constructor(
+        message: string,
+        errorCode: string = 'E2201',
+        httpResponseStatusCode: number = 500
+    ) {
+        super(message, errorCode, httpResponseStatusCode)
+    }
+}
