+++ src/WorkerProcess/AppScraperUtils/clearDir.ts
