+++ src/WorkerProcess/ScrapeConfigCache.ts
@@ -13,7 +13,7 @@ import type {
 } from '../types'
 import type { Authenticator } from './Authenticator'
 import { parseTemplate } from './htmlTemplates'
-import { getDashboardItemVisualization, getNestedPropertyValue } from './scrapeUtils'
+import { getDashboardItemVisualization, getNestedPropertyValue } from './AppScraperUtils'
 
 /* TODO: in the future type 'APP' should also be supported
  * But before we can do this we need a way to identify the correct URL
